nn working

+++ cases/trainings/ten_million/convert2binary.py
@@ -1,10 +1,10 @@
 import numpy as np
 
-exact_f = np.loadtxt('exact_f.dat')
-exact_angle = np.loadtxt('exact_angle.dat')
-exact_centroid = np.loadtxt('exact_centroid.dat')
-delta_angle = np.loadtxt('delta_angle.dat')
-initial_angle = np.loadtxt('initial_angle.dat')
+exact_f = np.loadtxt('exact_f.dat',dtype=np.float64)
+exact_angle = np.loadtxt('exact_angle.dat',dtype=np.float64)
+exact_centroid = np.loadtxt('exact_centroid.dat',dtype=np.float64)
+delta_angle = np.loadtxt('delta_angle.dat',dtype=np.float64)
+initial_angle = np.loadtxt('initial_angle.dat',dtype=np.float64)
 
 np.save('exact_f.npy',exact_f)
 np.save('exact_angle.npy',exact_angle)
